change how config is updated
The old config was parsed with a custom method. This commit replaces
that by some code from tornado. This is the same code that will be
used when the config is loaded in server.py.

Strings were not escaped when saving to config file. This won't
usually be a problem, but a malicious user may use that to circumvent
the `ALLOWED` check in ConfigHandler.post and probably more.