replace "[ ]" by checkbox on server to be more consistent with toggling
When toggling a checkbox, the client send the index of the checkbox
to the server. This means that server and client must exactly agree
on what is a checkbox and what isn't. Not suprisingly, this doesn't
work so well when they have different implementations. Furthermore,
the server uses the markdown file whereas the client only sees it
after it has been converted to HTML.

Markdown can handle HTML tags just fine, so we convert the "[ ]"
syntax to a checkbox before we pass the text to the markdown
converter.

I sneak in an additional change: I really like having checkboxes in
adjacent lines, but the old code (at least that in the browser)
required a paragraph for each checkbox. The new behavior is that of
the server and I add a <br/> to make it work well.

And one more: I use a <label> tag, so clicking on the text will
toggle the checkbox.